renamed service to repository mediator query code
