+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/AccountTransactions/GetAccountCoursePayments/GetAccountCoursePaymentsResponse.cs
@@ -1,9 +1,9 @@
 ﻿using System.Collections.Generic;
 using SFA.DAS.EAS.Domain.Models.Transaction;
 
-namespace SFA.DAS.EAS.Application.Queries.AccountTransactions.GetAccountCourseTransactions
+namespace SFA.DAS.EAS.Application.Queries.AccountTransactions.GetAccountCoursePayments
 {
-    public class GetAccountCourseTransactionsResponse
+    public class GetAccountCoursePaymentsResponse
     {
         public List<TransactionLine> Transactions { get; set; }
     }
